pyproject.toml
[tool.poetry]
name = "anisya"
version = "0.1.0"
description = ""
authors = ["wavy-cat <52543076+wavy-cat@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pynacl = "^1.5.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
